requirements.txt
google-genai
google-api-core
fastapi
pydantic>=2.0.0
uvicorn
python-multipart
httpx
tenacity
PyYAML
langchain-google-genai
pypdf2
langchain
python-docx
langchain-community
langchain-text-splitters
pinecone-client
rank_bm25
pytest
pytest-cov
pytest-asyncio
pytest-mock
cryptography
openai-whisper
setuptools-rust
